+++ packages/tts-editor/CHANGELOG.md
@@ -5,12 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [2.0.0] - 2025-03-02
 
 ### Breaking
 
-- Removed (undocumented) support for TSTL
-- Minimum VS Code version bumped to 1.96.0
+- Removed (undocumented) support for TSTL.
+- Minimum VS Code version bumped to 1.96.0.
 
 ### Changed